Merging stable bug-fixes into main

+++ ReporterDB.py
@@ -957,9 +957,8 @@ if __name__ == "__main__":
     input_path = args.input
     taxa_id = os.path.basename(input_path).split(".")[0]
 
-    print("Doing {}.".format(taxa_id))
-
     if 1 in verbose:
+        print("Doing {}.".format(taxa_id))
         T_init_db = time()
 
     if os.path.exists("/run/shm"):
@@ -1000,8 +999,10 @@ if __name__ == "__main__":
     nt_out_path = os.path.join(input_path, nt_out)
 
     if clear_output:
-        shutil.rmtree(aa_out_path)
-        shutil.rmtree(nt_out_path)
+        if os.path.exists(aa_out_path):
+            shutil.rmtree(aa_out_path)
+        if os.path.exists(nt_out_path):    
+            shutil.rmtree(nt_out_path)
 
     os.makedirs(aa_out_path, exist_ok=True)
     os.makedirs(nt_out_path, exist_ok=True)
